% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rwtkc_sum.R
\name{Rwtkc_sum}
\alias{Rwtkc_sum}
\title{Sum function for Rwtkc.}
\usage{
Rwtkc_sum(x, y)
}
\arguments{
\item{x}{x must be a numeric value, serving as the addend.}

\item{y}{y must be a numeric value, serving as the augend.}
}
\value{
The returned value is the sum of x and y.
}
\description{
This is an addition function developed by Rwtkc for testing purposes.
}
\examples{
Rwtkc_sum(1,2)
}
